Fix symbol IDs assigned to encoding packets
Previously, we incorrectly assigned internal symbol IDs (ISIs) to the
`PayloadId` of `EncodingPacket`s. According to RFC 6330, it should be
the encoding symbol IDs (ESIs). This commit fixes this inconsistency.

BREAKING CHANGE: As the assignment of symbol IDs changes, encoding
packets generated before this change cannot be decoded by the new
version and vice-versa.